backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.3
pymongo==4.5.0
pillow>=10.0.0
numpy>=1.24.0
tensorflow>=2.13.0
werkzeug>=2.3.0
python-dotenv>=1.0.0
redis>=4.6.0